+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
   [arsenico13]
 - **search-filters** endpoint now return also a list of searchable portal_types.
   [cekk]
+- Disable some types from *types_not_searched*.
+  [cekk]
 
 1.1.0 (2021-03-24)
 ------------------